+++ myems-api/app.py
@@ -294,6 +294,8 @@ api.add_route('/energystoragepowerstations/{id_}/export',
               energystoragepowerstation.EnergyStoragePowerStationExport())
 api.add_route('/energystoragepowerstations/import',
               energystoragepowerstation.EnergyStoragePowerStationImport())
+api.add_route('/energystoragepowerstations/{id_}/clone',
+              energystoragepowerstation.EnergyStoragePowerStationClone())
 
 
 api.add_route('/equipments',